requirements.txt
requests
jenkins-job-builder
PyYAML
responses
pytest
